Moving sandbox modules to a sub package
