+++ sandbox/src-js/example/Model.scala
@@ -2,6 +2,10 @@ package example
 
 import cats.effect.IO
 import tyrian.websocket.*
+import example.models.Fruit
+import example.models.Time
+import example.models.HttpDetails
+import example.models.Page
 
 import scala.scalajs.js
