+++ sandbox/src-js/example/models/Fruit.scala
@@ -1,3 +1,3 @@
-package example
+package example.models
 
 final case class Fruit(name: String, available: Boolean)
