+++ sandbox/src-js/example/models/Page.scala
@@ -1,4 +1,4 @@
-package example
+package example.models
 
 enum Page:
   case Page1, Page2, Page3, Page4, Page5, Page6, Page7
