+++ sandbox/src-js/example/models/Vector2.scala
@@ -1,3 +1,3 @@
-package example
+package example.models
 
 final case class Vector2(x: Double, y: Double)
